+++ src/index.ts
@@ -39,7 +39,6 @@ export type { AccessGrantAny } from "./type/AccessGrant";
 export type { AccessModes } from "./type/AccessModes";
 export type { FetchOptions } from "./type/FetchOptions";
 export type { RedirectOptions } from "./type/RedirectOptions";
-export type { SaveInContainerOptions } from "./type/SaveInContainerOptions";
 
 export { CredentialIsAccessGrantAny } from "./type/AccessGrant";